Move auto anchor generating to OTXSSD
